[Sprout Upstream Request] Fix touch error handling issue when S3203 touch boot up fail
[Fix Issue Description] In Sprout Factory test, if the phone haven’t use any touch panel,
the phone will boot up fail due to S3203 touch driver init time too long, this case may
lead to issues of phone always boot fail when you use another vendor’s touch panel on the s
ame device.

[Issue Root Cause] In S3203 driver, the i2c operation return value is set to unsigned char
type while the actual return type of i2c operation is int type, if the i2c operation returns
a negative value, the return value will be cut to a positive value due to unsigned char
return type, in this case, any transfer err will be treated as success, which will cause the
phone reboot in this case.

[Issue Fix] this change list is used to change unsigned char type to int type to fix this
issue.

+++ drivers/input/touchscreen/mediatek/S3203/synaptics_rmi4_driver.c
@@ -602,7 +602,7 @@ exit:
 int tpd_i2c_read_data_fifo(struct i2c_client *client,
                 unsigned short addr, unsigned char *data, unsigned short length)
 {
-        u8 retval=0;
+        int retval=0;
         u8 retry = 0;
         u8 *pData = data;
         int tmp_addr = addr;
@@ -727,7 +727,7 @@ exit:
 int tpd_i2c_write_data_fifo(struct i2c_client *client,
                 unsigned short addr, unsigned char *data, unsigned short length)
 {
-    u8 retval=0;
+    int retval=0;
     u8 retry = 0;
     u8 *pData = data;
     u8 buf[5] = {0};
